requirements.txt
fastapi
uvicorn
python-multipart
pillow
PyMuPDF
jinja2
